Crate libatm

libatm is a library for generating and working with MIDI files. It was purpose-built for All the Music, LLC to assist in its mission to enable musicians to make all of their music without the fear of frivolous copyright lawsuits. All code is released into the public domain via the Creative Commons Attribute 4.0 International License. If you're looking for a command line tool to generate and work with MIDI files, check out the atm-cli project that utilizes this library. For more information on All the Music, check out allthemusic.info.

Structs

MIDIChannelVoiceMessage
MIDI channel voice message

MIDIFile
MIDI file representation

MIDIHeader
MIDI file header

MIDINote
MIDI note

MIDINoteSequence
Container for sequence of MIDINote

MIDITrackHeader
MIDI track chunk header

Enums

MIDIFormat
MIDI file format

MIDINoteType
MIDI note type

MIDIStatus
MIDI message status